Addressed feedbacks from Pambo
